[IMP] Mail: Move suggestion code inside a new set of model/component
This will reduce composer model complexity and allow to reuse suggestion code
inside other component.

task-2447140
