+++ addons/mail/__manifest__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+
 {
     'name': 'Discuss',
     'version': '1.3',
@@ -115,8 +116,6 @@
             'mail/static/src/components/composer/composer.js',
             'mail/static/src/components/composer_suggested_recipient/composer_suggested_recipient.js',
             'mail/static/src/components/composer_suggested_recipient_list/composer_suggested_recipient_list.js',
-            'mail/static/src/components/composer_suggestion/composer_suggestion.js',
-            'mail/static/src/components/composer_suggestion_list/composer_suggestion_list.js',
             'mail/static/src/components/composer_text_input/composer_text_input.js',
             'mail/static/src/components/dialog/dialog.js',
             'mail/static/src/components/dialog_manager/dialog_manager.js',
@@ -149,6 +148,8 @@
             'mail/static/src/components/notification_popover/notification_popover.js',
             'mail/static/src/components/notification_request/notification_request.js',
             'mail/static/src/components/partner_im_status_icon/partner_im_status_icon.js',
+            'mail/static/src/components/suggestion/suggestion.js',
+            'mail/static/src/components/suggestions_list/suggestions_list.js',
             'mail/static/src/components/thread_icon/thread_icon.js',
             'mail/static/src/components/thread_needaction_preview/thread_needaction_preview.js',
             'mail/static/src/components/thread_preview/thread_preview.js',
@@ -192,6 +193,7 @@
             'mail/static/src/models/notification_group_manager/notification_group_manager.js',
             'mail/static/src/models/partner/partner.js',
             'mail/static/src/models/suggested_recipient_info/suggested_recipient_info.js',
+            'mail/static/src/models/suggestion_manager/suggestion_manager.js',
             'mail/static/src/models/thread/thread.js',
             'mail/static/src/models/thread_cache/thread_cache.js',
             'mail/static/src/models/thread_partner_seen_info/thread_partner_seen_info.js',
@@ -229,8 +231,8 @@
             'mail/static/src/components/composer/composer.scss',
             'mail/static/src/components/composer_suggested_recipient/composer_suggested_recipient.scss',
             'mail/static/src/components/composer_suggested_recipient_list/composer_suggested_recipient_list.scss',
-            'mail/static/src/components/composer_suggestion/composer_suggestion.scss',
-            'mail/static/src/components/composer_suggestion_list/composer_suggestion_list.scss',
+            'mail/static/src/components/suggestion/suggestion.scss',
+            'mail/static/src/components/suggestions_list/suggestions_list.scss',
             'mail/static/src/components/composer_text_input/composer_text_input.scss',
             'mail/static/src/components/dialog/dialog.scss',
             'mail/static/src/components/discuss/discuss.scss',
@@ -297,10 +299,10 @@
             'mail/static/src/components/chatter/chatter_suggested_recipient_tests.js',
             'mail/static/src/components/chatter_topbar/chatter_topbar_tests.js',
             'mail/static/src/components/composer/composer_tests.js',
-            'mail/static/src/components/composer_suggestion/composer_suggestion_canned_response_tests.js',
-            'mail/static/src/components/composer_suggestion/composer_suggestion_channel_tests.js',
-            'mail/static/src/components/composer_suggestion/composer_suggestion_command_tests.js',
-            'mail/static/src/components/composer_suggestion/composer_suggestion_partner_tests.js',
+            'mail/static/src/components/suggestions_list/suggestion_canned_response_tests.js',
+            'mail/static/src/components/suggestions_list/suggestion_channel_tests.js',
+            'mail/static/src/components/suggestions_list/suggestion_command_tests.js',
+            'mail/static/src/components/suggestions_list/suggestion_partner_tests.js',
             'mail/static/src/components/dialog_manager/dialog_manager_tests.js',
             'mail/static/src/components/discuss/tests/discuss_domain_tests.js',
             'mail/static/src/components/discuss/tests/discuss_inbox_tests.js',
@@ -375,8 +377,6 @@
             'mail/static/src/components/composer/composer.xml',
             'mail/static/src/components/composer_suggested_recipient/composer_suggested_recipient.xml',
             'mail/static/src/components/composer_suggested_recipient_list/composer_suggested_recipient_list.xml',
-            'mail/static/src/components/composer_suggestion/composer_suggestion.xml',
-            'mail/static/src/components/composer_suggestion_list/composer_suggestion_list.xml',
             'mail/static/src/components/composer_text_input/composer_text_input.xml',
             'mail/static/src/components/dialog/dialog.xml',
             'mail/static/src/components/dialog_manager/dialog_manager.xml',
@@ -409,6 +409,8 @@
             'mail/static/src/components/notification_popover/notification_popover.xml',
             'mail/static/src/components/notification_request/notification_request.xml',
             'mail/static/src/components/partner_im_status_icon/partner_im_status_icon.xml',
+            'mail/static/src/components/suggestions_list/suggestions_list.xml',
+            'mail/static/src/components/suggestion/suggestion.xml',
             'mail/static/src/components/thread_icon/thread_icon.xml',
             'mail/static/src/components/thread_needaction_preview/thread_needaction_preview.xml',
             'mail/static/src/components/thread_preview/thread_preview.xml',
