+++ addons/mail/static/src/components/composer_text_input/composer_text_input.xml
@@ -4,13 +4,13 @@
     <t t-name="mail.ComposerTextInput" owl="1">
         <div class="o_ComposerTextInput">
             <t t-if="composer">
-                <t t-if="composer.hasSuggestions">
-                    <ComposerSuggestionList
-                        composerLocalId="props.composerLocalId"
-                        isBelow="props.hasMentionSuggestionsBelowPosition"
-                    />
-                </t>
-                <textarea class="o_ComposerTextInput_textarea o_ComposerTextInput_textareaStyle" t-att-class="{ 'o-composer-is-compact': props.isCompact }" t-esc="composer.textInputContent" t-att-placeholder="textareaPlaceholder" t-on-click="_onClickTextarea" t-on-focusin="_onFocusinTextarea" t-on-focusout="_onFocusoutTextarea" t-on-keydown="_onKeydownTextarea" t-on-keyup="_onKeyupTextarea" t-on-input="_onInputTextarea" t-ref="textarea"/>
+                <SuggestionsList
+                    suggestionManagerLocalId="composer.suggestionManager.localId"
+                    isBelow="props.hasMentionSuggestionsBelowPosition"
+                    t-on-suggestion-selected="_onSuggestionSelected"
+                >
+                    <textarea class="o_ComposerTextInput_textarea o_ComposerTextInput_textareaStyle" t-att-class="{ 'o-composer-is-compact': props.isCompact }" t-esc="composer.textInputContent" t-att-placeholder="textareaPlaceholder" t-on-click="_onClickTextarea" t-on-focusin="_onFocusinTextarea" t-on-focusout="_onFocusoutTextarea" t-on-keydown="_onKeydownTextarea" t-on-keyup="_onKeyupTextarea" t-on-input="_onInputTextarea" t-ref="textarea"/>
+                </SuggestionsList>
                 <!--
                      This is an invisible textarea used to compute the composer
                      height based on the text content. We need it to downsize
