+++ addons/mail/static/src/components/suggestion/suggestion.scss
@@ -2,13 +2,13 @@
 // Layout
 // ------------------------------------------------------------------
 
-.o_ComposerSuggestion {
+.o_Suggestion {
     display: flex;
     width: map-get($sizes, 100);
     padding: map-get($spacers, 2) map-get($spacers, 4);
 }
 
-.o_ComposerSuggestion_part1 {
+.o_Suggestion_part1 {
     // avoid shrinking part 1 because it is more important than part 2
     // because no shrink, ensure it cannot overflow with a max-width
     flex: 0 0 auto;
@@ -18,14 +18,14 @@
     text-overflow: ellipsis;
 }
 
-.o_ComposerSuggestion_part2 {
+.o_Suggestion_part2 {
     // shrink part 2 to properly ensure it cannot overflow
     flex: 0 1 auto;
     overflow: hidden;
     text-overflow: ellipsis;
 }
 
-.o_ComposerSuggestion_partnerImStatusIcon {
+.o_Suggestion_partnerImStatusIcon {
     flex: 0 0 auto;
 }
 
@@ -33,11 +33,11 @@
 // Style
 // ------------------------------------------------------------------
 
-.o_ComposerSuggestion_part1 {
+.o_Suggestion_part1 {
     font-weight: $font-weight-bold;
 }
 
-.o_ComposerSuggestion_part2 {
+.o_Suggestion_part2 {
     font-style: italic;
     color: $gray-600;
 }
