+++ addons/mail/static/src/components/suggestion/suggestion.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates xml:space="preserve">
+
+    <t t-name="mail.Suggestion" owl="1">
+        <a class="o_Suggestion dropdown-item" t-att-class="{ 'active': props.isActive }" href="#" t-att-title="title()" role="menuitem" t-on-click="_onClick">
+            <t t-if="record">
+                <t t-if="isCannedResponse">
+                    <span class="o_Suggestion_part1"><t t-esc="record.source"/></span>
+                    <span class="o_Suggestion_part2"><t t-esc="record.substitution"/></span>
+                </t>
+                <t t-if="isChannel">
+                    <span class="o_Suggestion_part1"><t t-esc="record.name"/></span>
+                </t>
+                <t t-if="isCommand">
+                    <span class="o_Suggestion_part1"><t t-esc="record.name"/></span>
+                    <span class="o_Suggestion_part2"><t t-esc="record.help"/></span>
+                </t>
+                <t t-if="isPartner">
+                    <PartnerImStatusIcon
+                        class="o_Suggestion_partnerImStatusIcon"
+                        hasBackground="false"
+                        partnerLocalId="record.localId"
+                    />
+                    <span class="o_Suggestion_part1"><t t-esc="record.nameOrDisplayName"/></span>
+                    <t t-if="record.email">
+                        <span class="o_Suggestion_part2">(<t t-esc="record.email"/>)</span>
+                    </t>
+                </t>
+            </t>
+        </a>
+    </t>
+
+</templates>
