+++ addons/mail/static/src/components/suggestions_list/suggestion_canned_response_tests.js
@@ -1,8 +1,8 @@
-odoo.define('mail/static/src/components/composer_suggestion/composer_suggestion_canned_response_tests.js', function (require) {
+odoo.define('mail/static/src/components/suggestions_list/suggestion_canned_response_tests.js', function (require) {
 'use strict';
 
 const components = {
-    ComposerSuggestion: require('mail/static/src/components/composer_suggestion/composer_suggestion.js'),
+    ComposerSuggestion: require('mail/static/src/components/suggestion/suggestion.js'),
 };
 const {
     afterEach,
@@ -13,8 +13,8 @@ const {
 
 QUnit.module('mail', {}, function () {
 QUnit.module('components', {}, function () {
-QUnit.module('composer_suggestion', {}, function () {
-QUnit.module('composer_suggestion_canned_response_tests.js', {
+QUnit.module('suggestion', {}, function () {
+QUnit.module('suggestion_canned_response_tests.js', {
     beforeEach() {
         beforeEach(this);
 
@@ -53,7 +53,7 @@ QUnit.test('canned response suggestion displayed', async function (assert) {
         substitution: "Hello, how are you?",
     });
     await this.createComposerSuggestion({
-        composerLocalId: thread.composer.localId,
+        suggestionManagerLocalId: thread.composer.suggestionManager.localId,
         isActive: true,
         modelName: 'mail.canned_response',
         recordLocalId: cannedResponse.localId,
@@ -61,7 +61,7 @@ QUnit.test('canned response suggestion displayed', async function (assert) {
 
     assert.containsOnce(
         document.body,
-        `.o_ComposerSuggestion`,
+        `.o_Suggestion`,
         "Canned response suggestion should be present"
     );
 });
@@ -81,7 +81,7 @@ QUnit.test('canned response suggestion correct data', async function (assert) {
         substitution: "Hello, how are you?",
     });
     await this.createComposerSuggestion({
-        composerLocalId: thread.composer.localId,
+        suggestionManagerLocalId: thread.composer.suggestionManager.localId,
         isActive: true,
         modelName: 'mail.canned_response',
         recordLocalId: cannedResponse.localId,
@@ -89,26 +89,26 @@ QUnit.test('canned response suggestion correct data', async function (assert) {
 
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion',
+        '.o_Suggestion',
         "Canned response suggestion should be present"
     );
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion_part1',
+        '.o_Suggestion_part1',
         "Canned response source should be present"
     );
     assert.strictEqual(
-        document.querySelector(`.o_ComposerSuggestion_part1`).textContent,
+        document.querySelector(`.o_Suggestion_part1`).textContent,
         "hello",
         "Canned response source should be displayed"
     );
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion_part2',
+        '.o_Suggestion_part2',
         "Canned response substitution should be present"
     );
     assert.strictEqual(
-        document.querySelector(`.o_ComposerSuggestion_part2`).textContent,
+        document.querySelector(`.o_Suggestion_part2`).textContent,
         "Hello, how are you?",
         "Canned response substitution should be displayed"
     );
@@ -129,7 +129,7 @@ QUnit.test('canned response suggestion active', async function (assert) {
         substitution: "Hello, how are you?",
     });
     await this.createComposerSuggestion({
-        composerLocalId: thread.composer.localId,
+        suggestionManagerLocalId: thread.composer.suggestionManager.localId,
         isActive: true,
         modelName: 'mail.canned_response',
         recordLocalId: cannedResponse.localId,
@@ -137,11 +137,11 @@ QUnit.test('canned response suggestion active', async function (assert) {
 
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion',
+        '.o_Suggestion',
         "Canned response suggestion should be displayed"
     );
     assert.hasClass(
-        document.querySelector('.o_ComposerSuggestion'),
+        document.querySelector('.o_Suggestion'),
         'active',
         "should be active initially"
     );
