+++ addons/mail/static/src/components/suggestions_list/suggestion_partner_tests.js
@@ -1,8 +1,8 @@
-odoo.define('mail/static/src/components/composer_suggestion/composer_suggestion_partner_tests.js', function (require) {
+odoo.define('mail/static/src/components/suggestions_list/suggestion_partner_tests.js', function (require) {
 'use strict';
 
 const components = {
-    ComposerSuggestion: require('mail/static/src/components/composer_suggestion/composer_suggestion.js'),
+    ComposerSuggestion: require('mail/static/src/components/suggestion/suggestion.js'),
 };
 const {
     afterEach,
@@ -13,8 +13,8 @@ const {
 
 QUnit.module('mail', {}, function () {
 QUnit.module('components', {}, function () {
-QUnit.module('composer_suggestion', {}, function () {
-QUnit.module('composer_suggestion_partner_tests.js', {
+QUnit.module('suggestion', {}, function () {
+QUnit.module('suggestion_partner_tests.js', {
     beforeEach() {
         beforeEach(this);
 
@@ -53,7 +53,7 @@ QUnit.test('partner mention suggestion displayed', async function (assert) {
         name: "Demo User",
     });
     await this.createComposerSuggestion({
-        composerLocalId: thread.composer.localId,
+        suggestionManagerLocalId: thread.composer.suggestionManager.localId,
         isActive: true,
         modelName: 'mail.partner',
         recordLocalId: partner.localId,
@@ -61,7 +61,7 @@ QUnit.test('partner mention suggestion displayed', async function (assert) {
 
     assert.containsOnce(
         document.body,
-        `.o_ComposerSuggestion`,
+        `.o_Suggestion`,
         "Partner mention suggestion should be present"
     );
 });
@@ -82,7 +82,7 @@ QUnit.test('partner mention suggestion correct data', async function (assert) {
         name: "Demo User",
     });
     await this.createComposerSuggestion({
-        composerLocalId: thread.composer.localId,
+        suggestionManagerLocalId: thread.composer.suggestionManager.localId,
         isActive: true,
         modelName: 'mail.partner',
         recordLocalId: partner.localId,
@@ -90,7 +90,7 @@ QUnit.test('partner mention suggestion correct data', async function (assert) {
 
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion',
+        '.o_Suggestion',
         "Partner mention suggestion should be present"
     );
     assert.strictEqual(
@@ -100,21 +100,21 @@ QUnit.test('partner mention suggestion correct data', async function (assert) {
     );
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion_part1',
+        '.o_Suggestion_part1',
         "Partner's name should be present"
     );
     assert.strictEqual(
-        document.querySelector(`.o_ComposerSuggestion_part1`).textContent,
+        document.querySelector(`.o_Suggestion_part1`).textContent,
         "Demo User",
         "Partner's name should be displayed"
     );
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion_part2',
+        '.o_Suggestion_part2',
         "Partner's email should be present"
     );
     assert.strictEqual(
-        document.querySelector(`.o_ComposerSuggestion_part2`).textContent,
+        document.querySelector(`.o_Suggestion_part2`).textContent,
         "(demo_user@odoo.com)",
         "Partner's email should be displayed"
     );
@@ -135,7 +135,7 @@ QUnit.test('partner mention suggestion active', async function (assert) {
         name: "Demo User",
     });
     await this.createComposerSuggestion({
-        composerLocalId: thread.composer.localId,
+        suggestionManagerLocalId: thread.composer.suggestionManager.localId,
         isActive: true,
         modelName: 'mail.partner',
         recordLocalId: partner.localId,
@@ -143,11 +143,11 @@ QUnit.test('partner mention suggestion active', async function (assert) {
 
     assert.containsOnce(
         document.body,
-        '.o_ComposerSuggestion',
+        '.o_Suggestion',
         "Partner mention suggestion should be displayed"
     );
     assert.hasClass(
-        document.querySelector('.o_ComposerSuggestion'),
+        document.querySelector('.o_Suggestion'),
         'active',
         "should be active initially"
     );
