+++ addons/mail/static/src/components/suggestions_list/suggestions_list.scss
@@ -2,7 +2,7 @@
 // Layout
 // ------------------------------------------------------------------
 
-.o_ComposerSuggestionList {
+.o_SuggestionsList {
     position: absolute;
     // prevent suggestion items from overflowing
     width: 100%;
@@ -12,12 +12,12 @@
     }
 }
 
-.o_ComposerSuggestionList_drop {
+.o_SuggestionsList_drop {
     // prevent suggestion items from overflowing
     width: 100%;
 }
 
-.o_ComposerSuggestionList_list {
+.o_SuggestionsList_list {
     // prevent suggestion items from overflowing
     width: 100%;
     // prevent from overflowing chat window, must be smaller than its height
